Make the lasso tool (buried in an old experiment in story mapping) a bit more usable, and lift it to widgetry
